% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buzz1.R
\name{buzz1}
\alias{buzz1}
\title{Add 2 to a number or vector}
\usage{
buzz1(x)
}
\arguments{
\item{x}{A number or vector.}
}
\value{
\code{y} which is the sum of \code{x} and 2.
}
\description{
Add 2 to a number or vector
}
\examples{
buzz1(rnorm(5))
buzz1(3)
}
