% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytest.R
\name{mytest}
\alias{mytest}
\title{Produce the output from myone}
\usage{
mytest(n = 100)
}
\arguments{
\item{n}{An integer}
}
\value{
\code{x} which is a vector of random normal values,
\code{y} which is \code{x} + 4.
}
\description{
Produce the output from myone
}
